+++ database/migrations/2024_04_27_132810_create_product_prices_table.php
@@ -4,27 +4,26 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-public function up()
-{
-    Schema::create('product_prices', function (Blueprint $table) {
-        $table->id();
-        $table->float('list_price')->default(0);
-        $table->float('sale_price')->default(0);
-        $table->string('currency')->nullable();
-        $table->smallInteger('tax_rate')->nullable();
-        $table->unsignedBigInteger('product_id');
-        $table->unsignedBigInteger('price_field_id');
-        $table->timestamps();
+    public function up()
+    {
+        Schema::create('product_prices', function (Blueprint $table) {
+            $table->id();
+            $table->float('list_price')->default(0);
+            $table->float('sale_price')->default(0);
+            $table->string('currency')->nullable();
+            $table->float('tax_rate')->nullable();
+            $table->unsignedBigInteger('product_id');
+            $table->unsignedBigInteger('price_field_id');
+            $table->timestamps();
 
-        $table->foreign('product_id')->references('id')->on('products');
-        $table->foreign('price_field_id')->references('id')->on('price_fields');
-    });
-}
+            $table->foreign('product_id')->references('id')->on('products');
+            $table->foreign('price_field_id')->references('id')->on('price_fields');
+        });
+    }
 
 
     /**
